pyproject.toml
[tool.poetry]
name = "ot-harjoitustyo"
version = "0.1.0"
description = "WYSIWYG Editori Python"
authors = ["Lauri <1960498+msha@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
invoke = "^1.6.0"
tkhtmlview = "^0.1.0"
Pillow = "^8.4.0"
beautifulsoup4 = "^4.10.0"
dill = "^0.3.4"
lxml = "^4.7.1"
customtkinter = "^1.4"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
pytest = "^6.2.5"
coverage = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
